requirements.txt
requests>=2.31.0
supabase>=2.3.0
google-genai>=0.8.0
groq>=0.9.0
